Scaffold working properly

+++ src/tests/scaffold/test_scaffold.py
@@ -9,14 +9,16 @@ from src.enums import FileSystemEnum as En
 class TestScaffold(TestCase):
     def setUp(self) -> None:
         self.group = 'foo'
-        self.problem = 'bar.py'
+        self.problem = 'bar'
         self.scaffold = Scaffold(self.group, self.problem)
 
     def tearDown(self) -> None:
-        rmtree(self.group)
+        if Path(self.group).exists():
+            rmtree(self.group)
 
     def test_scaffold_WHEN_nothing_exists_THEN_creates_structure(self) -> None:
-        expected_file = f'{self.group}/{self.problem}/{En.KEY}/{En.KEY}{En.FILE_EXTENSION}'
+        expected_file = f'{self.group}/{self.problem}/{En.KEY}/{En.KEY}.{En.FILE_EXTENSION}'
+        print(expected_file)
 
         self.scaffold.create()
